requirements.txt
torch>=2.6.0
torchvision>=0.21.0
transformers>=4.50.0
Pillow>=12.0.0
PyAutoGUI>=0.9.54
timm>=1.0.0
bitsandbytes>=0.49.0
PyYAML>=6.0
Jinja2>=3.1.0
click>=8.1.0
colorama>=0.4.6
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
websockets>=14.0
